Match CSS files with the domain hostname
Load a CSS file based on the current page's domain (e.g. "example.com").

Use:
https://stackoverflow.com/questions/13367376/get-the-domain-name-of-the-subdomain-javascript/13367604#13367604
to give us the hostname without any subdomains. This will apply the CSS
to all subdomains on the site. Not sure if this what we want yet, but it
seemed simpler to consolidate like this now rather than splitting CSS by
each subdomain or adding some kind of complicated matching/wildcard
logic.

Errors are ignored and just happen if the file for the current domain
doesn't exist. This seems okay to me to leave errors unhandled, as there
doesn't appear to be a clean way to handle them, and even if there was,
we'd just be either ignoring them anyway or conditionally executing
`insertCSS()`, which seems unnecessary.