README: Add installation instructions
Includes instructions for Chrome and Firefox. Firefox is a little more
involved. For now not bothering to sign the extension.

Also update `manifest.json` to give the extension an add-on ID. This is
required as it must be used as the directory name in the Firefox
`extensions/` profile folder.

Thanks to these resources for explaining how to install an unpacked
extension in Firefox:
- https://stackoverflow.com/questions/27155766/how-to-install-unpacked-extension-in-firefox-chrome/27158189#27158189
- https://developer.mozilla.org/en-US/docs/Archive/Add-ons/Setting_up_extension_development_environment#Firefox_extension_proxy_file
- https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/Alternative_distribution_options/Sideloading_add-ons
- https://stackoverflow.com/questions/37728865/install-webextensions-on-firefox-from-the-command-line/37739112#37739112