Add a way to get rid of the reload keyboard shortcut for Chrome
The keyboard shortcut for reloading the extension causes the extension
to fail to load in Chrome. That happens because the default shortcut
uses the Control and Option modifiers together:

> Key combinations that involve Ctrl+Alt are not permitted in order to
> avoid conflicts with the AltGr key.

https://developer.chrome.com/docs/extensions/reference/commands/#key-combinations

Provide a way to remove the keyboard shortcuts from the manifest for
Chrome. The shortcuts aren't necessary in Chrome, as new CSS files and
changes to existing ones are automatically recognised without having to
reload the extension. On Firefox, the extension must be manually
reloaded in order for CSS changes to take effect.